+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [**BREAKING**] Upgraded gtk-rs to version 0.14. This is a breaking change
   because gtk-rs' API was changed in this release.
 
+### Added
+- `event_param` method for `woab::Signal`, to easily get the concrete GDK event
+  type from an event signal.
+
 ## 0.5.0 - 2021-05-19
 ### Added
 - `#[derive(woab::PropSync)]` for generating getter and setter for relevant